Fixed normalizing space in name_strings
When normalizing name strings we converted more than one space to single
space, and single space had beeen ignored. As a result names which had
a new line instead of space get stored in the database as is, which
caused problems during resolution. Now any single space is also
converted to 'normal' space.